test(plugin): cover fetch response size limits
